feat(nvim): use foot pid for hypr nav server

+++ dot_config/nvim/plugin/nvim-hypr-nav.lua
@@ -4,35 +4,66 @@ local M = {}
 -- Create a file $XDG_RUNTIME_DIR/nvim-hyp-nav.<pid>.servername that stores the
 -- server name
 local runtime_dir = vim.env.XDG_RUNTIME_DIR or "/tmp"
-local pid = tostring(vim.fn.getpid())
-M.servername_file = runtime_dir .. "/nvim-hypr-nav." .. pid .. ".servername"
+
+local function get_outermost_foot_pid()
+  if not vim.env.HYPRLAND_INSTANCE_SIGNATURE then
+    return nil
+  end
+
+  local pid = tostring(vim.fn.getpid())
+  local last_foot
+  while pid and pid ~= "1" do
+    local comm = vim.fn.systemlist({ "ps", "-o", "comm=", "-p", pid })[1]
+    comm = comm and vim.trim(comm)
+    if comm == "foot" or comm == "footclient" then
+      last_foot = pid
+    end
+    local ppid = vim.fn.systemlist({ "ps", "-o", "ppid=", "-p", pid })[1]
+    ppid = ppid and vim.trim(ppid)
+    pid = ppid
+  end
+  return last_foot
+end
+
+local server_pid = get_outermost_foot_pid()
+if server_pid then
+  M.servername_file = runtime_dir .. "/nvim-hypr-nav." .. server_pid .. ".servername"
+else
+  M.servername_file = nil
+end
 
 -- Write the servername file on VimEnter
 local augroup = vim.api.nvim_create_augroup("NvimHyprNav", { clear = true })
 vim.api.nvim_create_autocmd("VimEnter", {
-	group = augroup,
-	pattern = "*",
-	callback = function()
-		local servername = vim.v.servername
-		if servername ~= "" then
-			local file = io.open(M.servername_file, "w")
-			if not file then
-				vim.notify("Could not write servername file: " .. M.servername_file, vim.log.levels.ERROR)
-				return
-			end
-			file:write(servername)
-			file:close()
-		end
-	end,
+        group = augroup,
+        pattern = "*",
+        callback = function()
+                if not M.servername_file then
+                        return
+                end
+
+                local servername = vim.v.servername
+                if servername ~= "" then
+                        local file = io.open(M.servername_file, "w")
+                        if not file then
+                                vim.notify("Could not write servername file: " .. M.servername_file, vim.log.levels.ERROR)
+                                return
+                        end
+                        file:write(servername)
+                        file:close()
+                end
+        end,
 })
 
 -- Delete the servername file on exit
 vim.api.nvim_create_autocmd("VimLeavePre", {
-	group = augroup,
-	pattern = "*",
-	callback = function()
-		os.remove(M.servername_file)
-	end,
+        group = augroup,
+        pattern = "*",
+        callback = function()
+                if M.servername_file then
+                        os.remove(M.servername_file)
+                end
+        end,
 })
 
 -- Function to be called from the nvim_hypr_nav.sh script that lives in the
